folly/wangle/bootstrap/ServerBootstrap.h: avoid -Wsign-compare error (trivial)
Summary:
* folly/wangle/bootstrap/ServerBootstrap.h:
Change type of a for-loop index from int to size_t.

Test Plan:
Run this and note there are fewer errors than before:
fbconfig --platform-all=gcc-4.9-glibc-2.20 -r folly && fbmake dbgo

Reviewed By: jsedgwick@fb.com

Subscribers: fugalh, folly-diffs@

FB internal diff: D1770091

Tasks: 5941250

Signature: t1:1770091:1420667597:fbe9a5ba56a0e34e0480c82c8fb61128406ded95

+++ folly/wangle/bootstrap/ServerBootstrap.h
@@ -178,7 +178,7 @@ class ServerBootstrap {
     bind0->wait();
 
     auto barrier = std::make_shared<boost::barrier>(acceptor_group_->numThreads());
-    for (int i = 1; i < acceptor_group_->numThreads(); i++) {
+    for (size_t i = 1; i < acceptor_group_->numThreads(); i++) {
       acceptor_group_->add(std::bind(startupFunc, barrier));
     }
     barrier->wait();
